+++ system-parachains/collectives/collectives-polkadot/constants/src/lib.rs
@@ -27,3 +27,9 @@ pub const AMBASSADOR_SALARY_PALLET_INDEX: u8 = 74;
 
 /// Polkadot Ambassador Treasury pallet instance.
 pub const AMBASSADOR_TREASURY_PALLET_INDEX: u8 = 75;
+
+/// Polkadot Secretary Salary pallet instance.
+pub const SECRETARY_SALARY_PALLET_INDEX: u8 = 84;
+
+/// Polkadot Secretary Treasury pallet instance.
+pub const SECRETARY_TREASURY_PALLET_INDEX: u8 = 85;