+++ system-parachains/collectives/collectives-polkadot/src/fellowship/mod.rs
@@ -49,6 +49,10 @@ use sp_core::{ConstU128, ConstU32};
 use sp_runtime::traits::{ConstU16, ConvertToValue, IdentityLookup, Replace, TakeFirst};
 use xcm_builder::{AliasesIntoAccountId32, PayOverXcm};
 
+use crate::secretary::SecretaryCollectiveInstance;
+
+use crate::secretary::ranks::SECRETARY;
+
 #[cfg(feature = "runtime-benchmarks")]
 use crate::{
 	impls::benchmarks::{OpenHrmpChannel, PayWithEnsure},
@@ -84,19 +88,22 @@ impl pallet_referenda::Config<FellowshipReferendaInstance> for Runtime {
 	type RuntimeEvent = RuntimeEvent;
 	type Scheduler = Scheduler;
 	type Currency = Balances;
-	// Fellows can submit proposals.
+	// Fellows & Secretary can submit proposals.
 	type SubmitOrigin = EitherOf<
-		pallet_ranked_collective::EnsureMember<Runtime, FellowshipCollectiveInstance, 3>,
-		MapSuccess<
-			TryWithMorphedArg<
-				RuntimeOrigin,
-				<RuntimeOrigin as OriginTrait>::PalletsOrigin,
-				ToVoice,
-				EnsureOfRank<Runtime, FellowshipCollectiveInstance>,
-				(AccountId, u16),
+		EitherOf<
+			pallet_ranked_collective::EnsureMember<Runtime, FellowshipCollectiveInstance, 3>,
+			MapSuccess<
+				TryWithMorphedArg<
+					RuntimeOrigin,
+					<RuntimeOrigin as OriginTrait>::PalletsOrigin,
+					ToVoice,
+					EnsureOfRank<Runtime, FellowshipCollectiveInstance>,
+					(AccountId, u16),
+				>,
+				TakeFirst,
 			>,
-			TakeFirst,
 		>,
+		pallet_ranked_collective::EnsureMember<Runtime, SecretaryCollectiveInstance, { SECRETARY }>,
 	>;
 	type CancelOrigin = Architects;
 	type KillOrigin = Masters;